requirements.txt
rdkit>=2023.9.1
pillow>=10.0.0
